ca marche trop bien

+++ ajouterSalle.php
@@ -6,10 +6,21 @@
 
 if( isset( $_GET['proprioSalle'] ) && isset( $_GET['nomSalle'] ) && isset( $_GET['capaciteSalle'] ) && isset( $_GET['descriptionSalle'] )  ) {
 
-    $proprioSalle = $_GET['proprioSalle'];
-    $nomSalle = $_GET['nomSalle'];
-    $capaciteSalle = $_GET['capaciteSalle'];
-    $descriptionSalle = $_GET['descriptionSalle'];
+    ajouterNouvelleSalle( $_GET['proprioSalle'], $_GET['nomSalle'], $_GET['capaciteSalle'], $_GET['descriptionSalle'] );
+}
+else {
+
+    echo "Que fais tu ici ?";
+}
+
+/**
+ * @param $proprioSalle
+ * @param $nomSalle
+ * @param $capaciteSalle
+ * @param $descriptionSalle
+ */
+
+function ajouterNouvelleSalle( $proprioSalle, $nomSalle, $capaciteSalle, $descriptionSalle ) {
 
     if( !empty( $proprioSalle ) && !empty( $nomSalle ) && !empty( $capaciteSalle ) && !empty( $descriptionSalle ) ) {
 
@@ -17,24 +28,31 @@ if( isset( $_GET['proprioSalle'] ) && isset( $_GET['nomSalle'] ) && isset( $_GET
 
             $current_data = file_get_contents('data/data-test.json');
             $array_data = json_decode($current_data, true);
-            $form_data = array(
-                'nom' => $nomSalle,
-                'description' => $descriptionSalle,
-                'max-capacity' => $capaciteSalle,
-                'proprio' => $proprioSalle
 
-            );
+            if( !verifierSalleExistante($array_data['rooms'], $nomSalle ) ) {
 
-            //print_r($array_data);
-            $array_data['rooms'][] = $form_data;
-            $data_processed = json_encode($array_data, JSON_PRETTY_PRINT);
+                $form_data = array(
+                    'nom' => $nomSalle,
+                    'description' => $descriptionSalle,
+                    'max-capacity' => $capaciteSalle,
+                    'proprio' => $proprioSalle
 
-            if( file_put_contents( 'data/data-test.json', $data_processed ) ) {
+                );
 
-                echo "La liste des salles a été mise à jour.";
+                //print_r($array_data);
+                $array_data['rooms'][] = $form_data;
+                $data_processed = json_encode($array_data, JSON_PRETTY_PRINT);
+
+                if( file_put_contents( 'data/data-test.json', $data_processed ) ) {
+
+                    echo "La liste des salles a été mise à jour.";
+                }
+                else {
+                    echo "Il y a eu un problème pendant l'ouverture du fichier.";
+                }
             }
             else {
-                echo "Il y a eu un problème pendant l'ouverture du fichier.";
+                echo "Une salle portant ce nom existe déjà.";
             }
         }
         else {
@@ -45,6 +63,21 @@ if( isset( $_GET['proprioSalle'] ) && isset( $_GET['nomSalle'] ) && isset( $_GET
         echo "Les données du formulaire sont incomplètes.";
     }
 }
-else {
-    echo "Que fais tu ici ?";
+
+/**
+ * @param $array
+ * @param $nomSalle
+ * @return bool
+ */
+
+function verifierSalleExistante( $array, $nomSalle ) {
+
+    foreach ( $array as $a ) {
+
+        if( $a['nom'] == $nomSalle ) {
+            return true;
+        }
+    }
+
+    return false;
 }
